refactor: rename signer to signature submodule (#71)
